% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_size.R
\docType{data}
\name{cohort_size}
\alias{cohort_size}
\title{Tempcheck Cohort Sizes}
\format{
A data frame with 3 variables:

\describe{
}
}
\source{
{Customer Insight Team at NHS BSA}
}
\usage{
cohort_size
}
\description{
Tempcheck Cohort Sizes
}
\keyword{datasets}
