% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Tempcheck Data}
\format{
A data frame with ? variables:

\describe{
}
}
\source{
{Customer Insight Team at NHS BSA, via collection tempcheck_data in
tempcheck mongoDB database}
}
\usage{
data
}
\description{
Tempcheck Data
}
\keyword{datasets}
