% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_mongo_collection}
\alias{get_mongo_collection}
\title{Retrieve data from a MongoDB collection}
\usage{
get_mongo_collection(connection_string, db, collection)
}
\arguments{
\item{connection_string}{Address of the MongoDB server in mongo connection
string [URI
format](https://docs.mongodb.com/manual/reference/connection-string/).}

\item{db}{Name of the database the collection is in.}

\item{collection}{Name of the collection.}
}
\value{
Returns all data in the given collection as a dataframe.
}
\description{
Retrieve all data from a MongoDB collection, as a dataframe.
}
\examples{
\dontrun{
# Get the data in my_collection (in the my_db database)
get_mongo_collection(
  connection_string = glue::glue(
    "mongodb+srv://{MONGO_USER}:{MONGO_PASS}@{MONGO_HOST}",
    "/?tls=true&retryWrites=true&w=majority"
  ),
  "my_db",
  "my_collection"
)
}
}
