% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{send_to_mongo}
\alias{send_to_mongo}
\title{Append to or replace data in a MongoDB collection}
\usage{
send_to_mongo(data, connection_string, db, collection, replace = FALSE)
}
\arguments{
\item{connection_string}{Address of the MongoDB server in mongo connection
string [URI
format](https://docs.mongodb.com/manual/reference/connection-string/).}

\item{db}{Name of the database the collection is in.}

\item{collection}{Name of the collection.}

\item{replace}{Default is FALSE. When TRUE, all documents are dropped in the
collection first, effectively overwriting it.}
}
\value{
Returns all data in the given collection (after update) as a dataframe.
}
\description{
Append data to a MongoDB collection, optionally overwriting it.
  If the collection and/or database do not exist, they will be created.
}
\examples{
\dontrun{
# Append rows of my_df to my_collection (in the my_db database).
# Set `replace = FALSE` to instead overwrite it.
send_to_mongo(
  my_df,
  connection_string = glue::glue(
    "mongodb+srv://{MONGO_USER}:{MONGO_PASS}@{MONGO_HOST}",
    "/?tls=true&retryWrites=true&w=majority"
  ),
  "my_db",
  "my_collection",
  replace = FALSE
)
}
}
