% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{snap_to_df}
\alias{snap_to_df}
\title{Pull data from a SNAP survey}
\usage{
snap_to_df(snap_url, headers = character(0), query = character(0), col_names)
}
\arguments{
\item{snap_url}{URL of SNAP server}

\item{headers}{A named character vector of header keys and values. This allows
for the `X-USERNAME` and `X-API-KEY` headers to be set for SNAP  authentication.}

\item{query}{A named list of query params and values. This allows the
`restrictedValues` param to be set.}

\item{col_names}{A character vector of equal length to the number of variables
pulled from the survey. Columns of the resulting dataframe will be named from this.}
}
\value{
Returns a dataframe, with column names as specified in `col_names`.
}
\description{
Pull data from a SNAP survey. The data will be returned as
  a dataframe, with column names as specified.
}
\examples{
\dontrun{
# Get data from SNAP survey with ID 123-456, restricted to variables
# V18, V56, V58 and V49, as a dataframe. The columns will be named like
# V18 -> date, V56 -> team, V58 -> rating and V49 -> reason
snap_to_df(
  glue(SNAP_URL, SURVEY_ID = "123-456"),
  c(
    'X-USERNAME' = SNAP_USER,
    'X-API-KEY'  = SNAP_API_KEY
  ),
  list(restrictedVariables = "V18,V56,V58,V49"),
  col_names = c("date", "team", "rating", "reason")
)
}
}
