% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Tempcheck Data}
\format{
A data frame with 4 variables:

\describe{
}
}
\source{
{Customer Insight Team at NHS BSA, via SNAP survey with ID 55db4a62-7a8d-4df8-b8aa-0bf2c02a48a}
}
\usage{
data
}
\description{
Tempcheck Data
}
\keyword{datasets}
